--- src/operators/binary.rs
@@ -1,7 +1,0 @@
-use std::fmt::Debug;
-
-use crate::value::Value;
-
-pub trait BinaryOperator<V: Value>: Debug + Clone {
-    fn apply_2(&self, a: &V, b: &V) -> V;
-}
