--- src/operators/mod.rs
@@ -1,8 +1,0 @@
-pub mod binary;
-pub mod unary;
-
-#[derive(Clone, Debug)]
-pub enum Operator<TS: crate::TypeSystem> {
-    Binary(TS::BinaryOp),
-    Unary(TS::UnaryOp),
-}
